Modifying the EndOfEvent() method to fix a rare crash in the simulation of sds detectors

+++ sds/SdsMC/PndSdsDetector.cxx
@@ -210,7 +210,13 @@ void PndSdsDetector::EndOfEvent()
   if (fVerboseLevel)
     Print();
   
-  fPndSdsCollection->Delete();
+  if (fPndSdsCollection) {
+	  fPndSdsCollection->Delete();
+  } else {
+	  if (fVerboseLevel > 1) {
+		  std::cout << "-W- PndSdsDetector::EndOfEvent: no fPndSdsCollection pointer!" << std::endl;
+	  }
+  }
   fPosIndex = 0;
 }
 // -------------------------------------------------------------------------
